nghttpx: Fix Forwarded By
Previously, the value sent with Forward By parameter is the address
assigned to the listening socket.  This does not work expected if the
assigned address is wild card, (e.g., ::, 0.0.0.0).

This commit fixes this issue by using the local address that the
connection is accepted.  For QUIC, it is the local address that
handshake is performed, and it stays the same even after client
address migration occurred.