Cancel sending RST_STREAM if stream is not found
nghttp2_submit_rst_stream is intended to send RST_STREAM frame to the
existing stream.  Actually, nghttp2_session_add_rst_stream_continue
does not add RST_STREAM if the stream is not found.  There is a
situation that the stream exists when nghttp2_submit_rst_stream is
called, but it may be closed before actually sending RST_STREAM.
Previously, we send the frame in this case hoping that this is noop on
remote endpoint.  This commit checks stream existence just before
sending RST_STREAM, and if the stream is not found, cancel RST_STREAM.
This is the consistent behavior of nghttp2_submit_rst_stream and fixes
race condition.